+++ src/extension.ts
@@ -18,15 +18,16 @@
 */
 
 "use strict";
+
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
-// import * as vscode from 'vscode';
+// let fs = require("fs");
+import { window, workspace, commands, ExtensionContext, Disposable } from "vscode";
+import { basename, dirname, extname } from 'path';
 
 var init = false;
 var hasCpp = false;
 
-// let fs = require("fs");
-import { window, commands, ExtensionContext, Disposable } from "vscode";
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: ExtensionContext) {
@@ -90,7 +91,7 @@ export function activate(context: ExtensionContext) {
   commandArray.forEach((command) => {
     disposableCommandsArray.push(
       commands.registerCommand(command[0], () => {
-        commands.executeCommand(command[1]).then(function () {});
+        commands.executeCommand(command[1]).then(function () { });
       })
     );
   });
@@ -108,11 +109,11 @@ export function activate(context: ExtensionContext) {
       if (window.state.focused === true && !editor.selection.isEmpty) {
         commands
           .executeCommand("editor.action.formatSelection")
-          .then(function () {});
+          .then(function () { });
       } else {
         commands
           .executeCommand("editor.action.formatDocument")
-          .then(function () {});
+          .then(function () { });
       }
     }
   );
@@ -128,11 +129,11 @@ export function activate(context: ExtensionContext) {
       if (window.state.focused === true && !editor.selection.isEmpty) {
         commands
           .executeCommand("editor.action.formatSelection.multiple")
-          .then(function () {});
+          .then(function () { });
       } else {
         commands
           .executeCommand("editor.action.formatDocument.multiple")
-          .then(function () {});
+          .then(function () { });
       }
     }
   );
@@ -147,7 +148,7 @@ export function activate(context: ExtensionContext) {
       }
       commands
         .executeCommand("workbench.action.showAllEditorsByMostRecentlyUsed")
-        .then(function () {});
+        .then(function () { });
     }
   );
 
@@ -157,7 +158,7 @@ export function activate(context: ExtensionContext) {
       if (hasCpp) {
         commands
           .executeCommand("C_Cpp.SwitchHeaderSource")
-          .then(function () {});
+          .then(function () { });
       } else {
         window.showErrorMessage(
           "C/C++ extension (ms-vscode.cpptools) is not installed!"
@@ -179,8 +180,116 @@ export function activate(context: ExtensionContext) {
   context.subscriptions.push(disposableFormatWith);
   context.subscriptions.push(disposableSwitch);
 
-  //also update button in package.json.. see "Adding new buttons" in help.md file
+  // Adding 3 // user defined userButtons
+  for (let index = 1; index <= 10; index++) {
+    const printIndex = (index !== 10) ? ("0" + index) : ("" + index);
+    let action = "userButton" + printIndex;
+    let actionName = "ShortcutMenuBar." + action;
+    let disposableUserButtonCommand = commands.registerCommand(
+      actionName, () => {
+        const config = workspace.getConfiguration("ShortcutMenuBar");
+        let configName = action + "Command";
+        const command = config.get<String>(configName);
+
+        // skip userButtons not set
+        if (command === null || command === undefined || command.trimEnd() === "") {
+          return;
+        }
+
+
+        const palettes = command.split(',');
+        executeNext(action, palettes, 0);
+      }
+    );
+    context.subscriptions.push(disposableUserButtonCommand);
+  }
+
+  //also update userButton in package.json.. see "Adding new userButtons" in help.md file
 }
 
 // this method is called when your extension is deactivated
-export function deactivate() {}
+export function deactivate() { }
+
+
+// local functions for user-defined button execution follow, based on
+// https://github.com/ppatotski/vscode-commandbar/ Copyright 2018 Petr Patotski
+
+function executeNext(action: String, palettes: String[], index: number)  {
+  try {
+    let [cmd, ...args] = palettes[index].split('|');
+    if (args) {
+      args = args.map(arg => resolveVariables(arg));
+    }
+    cmd = cmd.trim();
+    commands.executeCommand(cmd, ...args).then(() => {
+      index++;
+      if (index < palettes.length) {
+        executeNext(action, palettes, index);
+      }
+    }, (err) => {
+      window.showErrorMessage(`Execution of '${action}' command has failed: ${err.message}`);
+    });
+  } catch (err) {
+    window.showErrorMessage(`Execution of '${action}' command has failed: ${err.message}`);
+    console.error(err);
+  }
+};
+
+const resolveVariablesFunctions = {
+  env: (name) => process.env[name.toUpperCase()],
+  cwd: () => process.cwd(),
+  workspaceRoot: () => getWorkspaceFolder(),
+  workspaceFolder: () => getWorkspaceFolder(),
+  workspaceRootFolderName: () => basename(getWorkspaceFolder()),
+  workspaceFolderBasename: () => basename(getWorkspaceFolder()),
+  lineNumber: () => window.activeTextEditor?.selection.active.line,
+  selectedText: () => window.activeTextEditor?.document.getText(window.activeTextEditor.selection),
+  file: () => getActiveEditorName(),
+  fileDirname: () => dirname(getActiveEditorName()),
+  fileExtname: () => extname(getActiveEditorName()),
+  fileBasename: () => basename(getActiveEditorName()),
+  fileBasenameNoExtension: () => {
+    const edtBasename = basename(getActiveEditorName());
+    return edtBasename.slice(0, edtBasename.length - extname(edtBasename).length);
+  },
+  execPath: () => process.execPath
+};
+
+const variableRegEx = /\$\{(.*?)\}/g;
+function resolveVariables(commandLine: String) {
+  return commandLine.trim().replace(variableRegEx, function replaceVariable(match, variableValue) {
+    const [variable, argument] = variableValue.split(':');
+    const resolver = resolveVariablesFunctions[variable];
+    if (!resolver) {
+      throw new Error(`Variable ${variable} not found!`);
+    }
+
+    return resolver(argument);
+  });
+};
+
+function getActiveEditorName() {
+  if (window.activeTextEditor) {
+    return window.activeTextEditor.document.fileName;
+  }
+  return "";
+}
+
+function getWorkspaceFolder(activeTextEditor = window.activeTextEditor) {
+  let folder;
+  if (workspace?.workspaceFolders) {
+    if (workspace.workspaceFolders.length === 1) {
+      folder = workspace.workspaceFolders[0].uri.fsPath;
+    } else if (activeTextEditor) {
+      const folderObject = workspace.getWorkspaceFolder(activeTextEditor.document.uri);
+      if (folderObject) {
+        folder = folderObject.uri.fsPath;
+      } else {
+        folder = "";
+      }
+    } else if (workspace.workspaceFolders.length > 0) {
+      folder = workspace.workspaceFolders[0].uri.fsPath;
+    }
+  }
+  return folder;
+}